requirements.txt
google-adk
python-dotenv
google-generativeai
litellm
yfinance
psutil
